requirements.txt
flask
requests
gunicorn
werkzeug
flask-session
python-dotenv